Add files via upload
In this problem I had to find the shortest cycle in a graph. Shortest cycle in a graph can be found by using BFS and keeping track of each nodes parents. We need to take all the nodes as our root and find if there is a cycle involving that node as root. If so , which cycle path is minimum.